[RTL] - Fix a number of MSVC/64 bit warnings/problems - Fix return type of RtlpInitializeHeapSegment (by Roel)
svn path=/trunk/; revision=53764

+++ reactos/lib/rtl/rtlp.h
@@ -63,7 +63,7 @@ typedef struct _DISPATCHER_CONTEXT
 PVOID
 NTAPI
 RtlpAllocateMemory(
-    ULONG Bytes,
+    SIZE_T Bytes,
     ULONG Tag);
 
 VOID
